+++ config.sample.json
@@ -14,6 +14,7 @@
 			"pass": ""
 		}
 	},
+	"debugEmail": true,
 
 	"port": 3001,
 	"cookie_secret": "I return from whence I came...",
