+++ src/activate.js
@@ -46,27 +46,9 @@ function passwordReset(req,res,next)
 			}
 			else
 			{
-				// build email message
-				var template = jade.compile( fs.readFileSync(libpath.resolve(__dirname, '../templates/activate-email.jade')) );
-				var html = template({
-					registration: false,
-					url: config.origin,
-					token: token,
-					username: req.body.username
-				});
-
-				// send out confirmation email
 				global.log('Sent out password token:', token);
-				var transporter = nodemailer.createTransport(config.smtp);
-				transporter.sendMail({
-					from: 'no-reply@toonstore.net',
-					to: req.body.email,
-					subject: 'Reset your password - ToonStore.net',
-					html: html
-				});
-								
+				global.renderActivationEmail(req.body.email, req.body.username, token, false, res.i18n.__.bind(res.i18n));
 				res.redirect('/pre-activate');
-
 			}
 			connection.end();
 		}
