+++ src/global.js
@@ -1,8 +1,13 @@
 /*
  * Stores global functions and handlers that are always useful
  */
-var libpath = require('path'),
-	moment = require('moment');
+var jade = require('jade'),
+	fs = require('fs'),
+	libpath = require('path'),
+	nodemailer = require('nodemailer'),
+	moment = require('moment'),
+
+	config = require('../config.json');
 
 var logLevels = {
 	'fatal': 0,
@@ -127,9 +132,37 @@ function renderPage(template, options)
 	return middleware;
 }
 
+function renderActivationEmail(to, username, token, registering, __)
+{
+	// build email message
+	var template = jade.compile( fs.readFileSync(libpath.resolve(__dirname, '../templates/activate-email.jade')), {pretty: true} );
+	var html = template({
+		registration: registering,
+		url: config.origin,
+		token: token,
+		username: username,
+		__: __
+	});
+
+	// send out confirmation email
+	if(!config.debugEmail){
+		var transporter = nodemailer.createTransport(config.smtp);
+		transporter.sendMail({
+			from: 'no-reply@toonstore.net',
+			to: to,
+			subject: (registering ? 'Set' : 'Reset') + ' your password - ToonStore.net',
+			html: html
+		});
+	}
+	else {
+		log((registering ? 'Set' : 'Reset') + ' your password - ToonStore.net');
+		log(html);
+	}
+}
+
 // export everything for external modules
 exports.error = error;
 exports.log = log;
 exports.logLevels = logLevels;
 exports.renderPage = renderPage;
-
+exports.renderActivationEmail = renderActivationEmail;
